+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 ## Unreleased
 
-- Memoize calls to `get_model()` so that model instances with the same parameters are cached and re-used (pass `use_cache=False` to disable).
+- Memoize calls to `get_model()` so that model instances with the same parameters are cached and re-used (pass `memoize=False` to disable).
 - Async context manager for `Model` class for optional scoped usage of model clients.
 
 ## v0.3.64 (14 February 2025)